refactor transformToRequire url handling
