+++ lib/template-compiler/modules/transform-srcset.js
@@ -1,5 +1,7 @@
 // vue compiler module for transforming `img:srcset` to a number of `require`s
 
+const urlToRequire = require('../url-to-require')
+
 module.exports = () => ({
   postTransformNode: node => {
     transform(node)
@@ -41,19 +43,5 @@ function transform (node) {
         attr.value = code
       }
     })
-  }
-}
-
-function urlToRequire (url) {
-  // same logic as in transform-require.js
-  const firstChar = url.charAt(0)
-  if (firstChar === '.' || firstChar === '~') {
-    if (firstChar === '~') {
-      const secondChar = url.charAt(1)
-      url = url.slice(secondChar === '/' ? 2 : 1)
-    }
-    return `require("${url}")`
-  } else {
-    return `"${url}"`
   }
 }
